requirements.txt
python-dotenv
pandas
selenium
bs4
pyodbc
firebase_admin